@@ -17,7 +17,7 @@
     "types/index.d.ts"
   ],
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "scripts": {
     "lint": "eslint --ignore-path .gitignore .",
@@ -26,13 +26,13 @@
     "test:types": "tsd"
   },
   "dependencies": {
-    "@percy/sdk-utils": "~1.0.0-beta.44"
+    "@percy/sdk-utils": "^1.0.0"
   },
   "peerDependencies": {
     "webdriverio": "~6 || ~7"
   },
   "devDependencies": {
-    "@percy/core": "~1.0.0-beta.44",
+    "@percy/core": "^1.0.0",
     "@wdio/cli": "^7.0.8",
     "@wdio/local-runner": "^7.0.0",
     "@wdio/mocha-framework": "^7.0.7",
